LiquidTest handler class
